+++ src/pbrt/base/camera.h
@@ -52,9 +52,9 @@ struct CameraTransform {
 struct CameraSample {
     Point2f p_film;
     Point2f p_lens;
-    double filter_weight = 1;
+    FloatType filter_weight = 1;
 
-    PBRT_GPU CameraSample(const Point2f _p_film, const Point2f _p_lens, double _filter_weight)
+    PBRT_GPU CameraSample(const Point2f _p_film, const Point2f _p_lens, FloatType _filter_weight)
         : p_film(_p_film), p_lens(_p_lens), filter_weight(_filter_weight) {}
 };
