+++ src/pbrt/samplers/independent.h
@@ -11,7 +11,7 @@ class IndependentSampler : public Sampler {
         curand_init(seed, 0, 0, &rand_state);
     }
 
-    PBRT_GPU double get_1d() override {
+    PBRT_GPU FloatType get_1d() override {
         return curand_uniform(&rand_state);
     }
